+++ example/plugins/echo.py
@@ -5,5 +5,5 @@ plugin = Plugin(name="Echo")
 
 
 @plugin.on_startswith_text("echo")
-async def on_echo(message, env, body):
-    await env.reply("{}".format(body), attachment=message.attachments)
+async def on_echo(message, env):
+    await env.reply("{}".format(env.body), attachment=message.attachments)