+++ test/test_manager_debug.py
@@ -7,9 +7,9 @@ class TestManagerDebug(KutanaTest):
 
         with self.debug_manager(["echo message"]) as plugin:
 
-            async def on_echo(message, env, body):
+            async def on_echo(message, env):
                 self.assertIsNone(await env.request("users.get"))
-                await env.reply(body)
+                await env.reply(env.body)
 
             plugin.on_startswith_text("echo")(on_echo)
 
@@ -32,8 +32,8 @@ class TestManagerDebug(KutanaTest):
         self.target = ["message"]
 
         with self.debug_manager(["echo message"]) as plugin:
-            async def on_echo(message, env, body):
-                await env.send_message(body, 1)
+            async def on_echo(message, env):
+                await env.send_message(env.body, 1)
 
             plugin.on_startswith_text("echo")(on_echo)
 
@@ -47,8 +47,8 @@ class TestManagerDebug(KutanaTest):
                 (3, "echo message4")
         ]) as plugin:
 
-            async def on_echo(message, env, body):
-                await env.reply(body)
+            async def on_echo(message, env):
+                await env.reply(env.body)
 
             plugin.on_startswith_text("echo")(on_echo)