+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     module-level variables.
   - ^ Only one dispose or startup callback can be present on one plugin.
   - ^ Executor's `register*` methods no longer returns decorators.
+  - ^ Plugin's arguments like "body", "args" e.t.c. now stored in environment.
   - Renamed Kutana's "process_update" to "process"
   - Processing is now done inside on Environment
   - Callbacks storing changed